feat: 6단계 코드 수정 성공
