+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,10 +1,15 @@
 package roomescape.controller;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
 import org.springframework.web.bind.annotation.*;
 import roomescape.dto.Reservation;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -16,10 +21,42 @@ public class ReservationController
     private List<Reservation> reservations = new ArrayList<>(); // 예약 정보들이 들어있는 리스트
     private AtomicLong index = new AtomicLong(1); // 예약자 번호
 
+
+    // 6단계 : JdbcTemplate을 이용하여 DataSource객체에 접근
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+
+    // ReservationRowmapper 클래스가 RowMapper 인터페이스를 구현한다.
+    public class ReservationRowmapper implements RowMapper<Reservation> {
+        @Override
+        public Reservation mapRow(ResultSet rs, int rowNum) throws SQLException {
+            Long id = rs.getLong("id");
+            String name = rs.getString("name");
+            String date = rs.getString("date");
+            String time = rs.getString("time");
+
+            return new Reservation(id, name, date, time);
+        }
+    }
+
+    // 6단계 : 데이터 조회하기 : 전체 데이터 조회하기
+    public List<Reservation> getReservationByDB() {
+        String sql = "SELECT * FROM reservation";
+        return jdbcTemplate.query(sql, new ReservationRowmapper());
+    }
+
+
+    // 6단계 : 데이터 조회하기 : 특정 데이터 조회하기 (id 활용)
+    public Reservation getReservationByDB(Long id) {
+        String sql = "SELECT * FROM reservation WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql, new ReservationRowmapper(), id);
+    }
+
+
     // localhost:8080/reservations 요청 시 예약자 현황 리스트가 응답할 수 있도록 구현
     @GetMapping("/reservations")
     public List<Reservation> getReservations() {
-        return reservations;
+        return getReservationByDB();
     }
 
     // 예약 추가
