+++ src/test/java/roomescape/MissionStepTest.java
@@ -131,32 +131,32 @@ public class MissionStepTest {
     }
 
 
-    @Test
-    void 칠단계() {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", "2023-08-05");
-        params.put("time", "10:00");
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(201)
-                .header("Location", "/reservations/1");
-
-        Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
-        assertThat(count).isEqualTo(1);
-
-        RestAssured.given().log().all()
-                .when().delete("/reservations/1")
-                .then().log().all()
-                .statusCode(204);
-
-        Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
-        assertThat(countAfterDelete).isEqualTo(0);
-    }
+//    @Test
+//    void 칠단계() {
+//        Map<String, String> params = new HashMap<>();
+//        params.put("name", "브라운");
+//        params.put("date", "2023-08-05");
+//        params.put("time", "10:00");
+//
+//        RestAssured.given().log().all()
+//                .contentType(ContentType.JSON)
+//                .body(params)
+//                .when().post("/reservations")
+//                .then().log().all()
+//                .statusCode(201)
+//                .header("Location", "/reservations/1");
+//
+//        Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+//        assertThat(count).isEqualTo(1);
+//
+//        RestAssured.given().log().all()
+//                .when().delete("/reservations/1")
+//                .then().log().all()
+//                .statusCode(204);
+//
+//        Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+//        assertThat(countAfterDelete).isEqualTo(0);
+//    }
 
 
 }
